Home Activity reminder code to Reminder Fragment
changes all reminder checking to be done in Reminder Fragment
bug fix - when editing date based reminders, the previously set date
will display properly
bug fix - editing date based reminders and saving without making changes
will no longer make the date unparseable
